pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "oviqs"
version = "0.1.0"
description = "OpenVINO Inference Quality Suite for LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "OVIQS contributors" }]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pydantic>=2.6",
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
  "pandas>=2.0",
]

[project.optional-dependencies]
openvino = [
  "openvino>=2025.0",
  "optimum-intel[openvino]>=1.20",
  "transformers>=4.40",
  "torch>=2.2",
]
genai = ["openvino-genai>=2025.0"]
eval = ["lighteval>=0.8", "lm-eval[api]>=0.4.0"]
rag = ["ragas>=0.2"]
agent = ["deepeval>=2.0", "openevals>=0.0.1", "agentevals>=0.0.1"]
observability = ["arize-phoenix>=7.0", "opik>=1.0"]
plots = ["matplotlib>=3.8"]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
  "build>=1.2",
  "twine>=6.0",
  "bandit>=1.8",
  "pip-audit>=2.8",
  "pre-commit>=4.0",
  "pyright>=1.1",
  "deptry>=0.23",
  "codespell>=2.3",
  "interrogate>=1.7",
  "pydoclint>=0.6",
  "vulture>=2.14",
  "radon>=6.0",
  "xenon>=0.9",
  "uv>=0.5",
]

[project.scripts]
oviq = "oviqs.cli:app"

[project.entry-points."oviqs.runners"]
dummy = "oviqs.adapters.runners.dummy:DummyLogitsRunner"

[project.entry-points."oviqs.datasets"]
jsonl = "oviqs.adapters.datasets:JsonlDatasetAdapter"

[project.entry-points."oviqs.reporters"]
json = "oviqs.adapters.reporting:JsonReportAdapter"
markdown = "oviqs.adapters.reporting:MarkdownReportRenderer"

[tool.hatch.build.targets.wheel]
packages = ["src/oviqs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "openvino: tests requiring OpenVINO runtime or exported models",
  "integration: integration tests",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv", "build", "dist"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = false
reportMissingTypeStubs = false
reportPrivateImportUsage = false

[tool.coverage.run]
source = ["oviqs"]

[tool.coverage.report]
fail_under = 60
show_missing = true
skip_empty = true

[tool.deptry]
extend_exclude = ["build", "dist", "reports", "models", "data", "datasets"]
known_first_party = ["oviqs"]
optional_dependencies_dev_groups = ["dev"]
package_module_name_map = { "optimum-intel" = ["optimum"] }

[tool.deptry.per_rule_ignores]
DEP001 = [
  "datasets",
  "opencompass",
  "phoenix",
]
DEP002 = [
  "agentevals",
  "arize-phoenix",
  "matplotlib",
  "pandas",
]

[tool.codespell]
skip = ".git,.venv,build,dist,reports,models,cache,data,datasets,artifacts,outputs,logs,*.egg-info"
ignore-words-list = "nd,ot,rouge"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-module = true
ignore-nested-functions = true
ignore-private = true
fail-under = 0
exclude = ["build", "dist", ".venv"]

[tool.pydoclint]
style = "google"
exclude = '\.venv|build|dist'
allow-init-docstring = true
check-return-types = false
check-yield-types = false

[tool.vulture]
paths = ["src", "tests"]
min_confidence = 100
sort_by_size = true
exclude = ["build", "dist", ".venv"]
